测试步进电机以及步进电机驱动ic 结论: 1:步进电机驱动IC clk有输入频率上限,当前ic为THB6128 上限为28khz 2:步进电机(42步进电机 step1.8°)自启动最大频率大约1khz Signed-off-by: gepingchao <gpc248038291912@163.com>

+++ Src/motor.h
@@ -41,7 +41,7 @@ typedef struct
 	unsigned int timer_value;
 	
 	unsigned int adjust;
-	float speed;
+	int speed;
 	float current_position;
 	float current_angle;
 	unsigned int distance;
